Use OverlayStore as Snapshot; define the interface of Snapshot
