+++ crates/store/src/lib.rs
@@ -1,9 +1,8 @@
 mod db_utils;
-mod overlay;
 mod snapshot;
 mod store_impl;
 pub mod transaction;
 mod write_batch;
 
-pub use overlay::OverlayStore;
+pub use snapshot::Snapshot;
 pub use store_impl::Store;
